New Readme + Folders/filenames refactoring
