+++ README.md
@@ -1,1 +1,28 @@
 # LEFTNIX v0.1.0
+
+## How commie is your favorite movie?
+
+_LEFTNIX is a movie rating platform that allows the audience to see though the flashy eye-candy million dollar money laundering schemes everyone insists to call movies... and I kind of forgot how the sentence started._
+_Wait... Oh, right: -to see through the movies and find out what ethics and political ideologies they present through their narative._
+
+> Are you ready to uncover the truth?!
+
+---
+
+## Up next...
+
+- TOP 3 Commie Movies
+- Movie Trailer Modal
+
+### And then...
+
+- Other pages besides the Gallery
+  \*\* Also User profile
+- Login splash/page _(probably fake)_
+
+---
+
+#### Wait, what?
+
+LEFTNIX is a joke and it doesn't _work_ - yet.
+But this joke is also a front-end practice which fetches movie data from [TMDB](https://www.themoviedb.org 'The Movie Database Website') API. :)
