+++ src/components/MovieGallery/index.jsx
@@ -1,10 +1,10 @@
-import MovieItem from './MovieItem';
+import MovieCard from '../MovieCard';
 
 export default function MovieGallery({ movies }) {
 	return (
 		<div id="movie-gallery-wrapper">
 			{movies.map((movie) => {
-				return <MovieItem key={movie.id} movie={movie} />;
+				return <MovieCard key={movie.id} movie={movie} />;
 			})}
 		</div>
 	);
